setup k6 for typescript
